+++ Fwk/AppFwk/CommonCode/cvfStructGridGeometryGenerator.cpp
@@ -57,114 +57,6 @@ namespace cvf
 ///
 //==================================================================================================
 
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-CellRangeFilter::CellRangeFilter()
-{
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void CellRangeFilter::addCellIncludeRange( size_t minI,
-                                           size_t minJ,
-                                           size_t minK,
-                                           size_t maxI,
-                                           size_t maxJ,
-                                           size_t maxK,
-                                           bool   applyToSubGridAreas )
-{
-    m_includeRanges.push_back( CellRange( minI, minJ, minK, maxI, maxJ, maxK, applyToSubGridAreas ) );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void CellRangeFilter::addCellInclude( size_t i, size_t j, size_t k, bool applyToSubGridAreas )
-{
-    m_includeRanges.push_back( CellRange( i, j, k, i + 1, j + 1, k + 1, applyToSubGridAreas ) );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void CellRangeFilter::addCellExcludeRange( size_t minI,
-                                           size_t minJ,
-                                           size_t minK,
-                                           size_t maxI,
-                                           size_t maxJ,
-                                           size_t maxK,
-                                           bool   applyToSubGridAreas )
-{
-    m_excludeRanges.push_back( CellRange( minI, minJ, minK, maxI, maxJ, maxK, applyToSubGridAreas ) );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-void CellRangeFilter::addCellExclude( size_t i, size_t j, size_t k, bool applyToSubGridAreas )
-{
-    m_excludeRanges.push_back( CellRange( i, j, k, i + 1, j + 1, k + 1, applyToSubGridAreas ) );
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-bool CellRangeFilter::isCellVisible( size_t i, size_t j, size_t k, bool isInSubGridArea ) const
-{
-    if ( m_includeRanges.empty() )
-    {
-        return false;
-    }
-
-    size_t idx;
-    for ( idx = 0; idx < m_excludeRanges.size(); idx++ )
-    {
-        if ( m_excludeRanges[idx].isInRange( i, j, k, isInSubGridArea ) )
-        {
-            return false;
-        }
-    }
-
-    for ( idx = 0; idx < m_includeRanges.size(); idx++ )
-    {
-        if ( m_includeRanges[idx].isInRange( i, j, k, isInSubGridArea ) )
-        {
-            return true;
-        }
-    }
-
-    return false;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-bool CellRangeFilter::isCellExcluded( size_t i, size_t j, size_t k, bool isInSubGridArea ) const
-{
-    for ( size_t idx = 0; idx < m_excludeRanges.size(); idx++ )
-    {
-        if ( m_excludeRanges[idx].isInRange( i, j, k, isInSubGridArea ) )
-        {
-            return true;
-        }
-    }
-
-    return false;
-}
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
-bool CellRangeFilter::hasIncludeRanges() const
-{
-    if ( !m_includeRanges.empty() )
-        return true;
-    else
-        return false;
-}
-
 //==================================================================================================
 ///
 /// \class cvf::StructGridGeometry
