+++ Fwk/AppFwk/CommonCode/cvfStructGridGeometryGenerator.h
@@ -50,71 +50,6 @@ class StructGridScalarDataAccess;
 //--------------------------------------------------------------------------------------------------
 ///
 //--------------------------------------------------------------------------------------------------
-class CellRangeFilter
-{
-public:
-    CellRangeFilter();
-
-    void addCellIncludeRange( size_t minI, size_t minJ, size_t minK, size_t maxI, size_t maxJ, size_t maxK, bool applyToSubGridAreas );
-    void addCellInclude( size_t i, size_t j, size_t k, bool applyToSubGridAreas );
-
-    void addCellExcludeRange( size_t minI, size_t minJ, size_t minK, size_t maxI, size_t maxJ, size_t maxK, bool applyToSubGridAreas );
-    void addCellExclude( size_t i, size_t j, size_t k, bool applyToSubGridAreas );
-
-    bool isCellVisible( size_t i, size_t j, size_t k, bool isInSubGridArea ) const;
-    bool isCellExcluded( size_t i, size_t j, size_t k, bool isInSubGridArea ) const;
-
-    bool hasIncludeRanges() const;
-
-private:
-    class CellRange
-    {
-    public:
-        CellRange()
-            : m_min( cvf::Vec3st::ZERO )
-            , m_max( UNDEFINED_SIZE_T, UNDEFINED_SIZE_T, UNDEFINED_SIZE_T )
-            , m_applyToSubGridAreas( true )
-        {
-        }
-
-        CellRange( size_t minI, size_t minJ, size_t minK, size_t maxI, size_t maxJ, size_t maxK, bool applyToSubGridAreas )
-            : m_min( minI, minJ, minK )
-            , m_max( maxI, maxJ, maxK )
-            , m_applyToSubGridAreas( applyToSubGridAreas )
-        {
-        }
-
-        bool isInRange( size_t i, size_t j, size_t k, bool isInSubGridArea ) const
-        {
-            if ( isInSubGridArea && !m_applyToSubGridAreas ) return false;
-            cvf::Vec3st test( i, j, k );
-
-            int idx;
-            for ( idx = 0; idx < 3; idx++ )
-            {
-                if ( test[idx] < m_min[idx] || m_max[idx] <= test[idx] )
-                {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
-    public:
-        cvf::Vec3st m_min;
-        cvf::Vec3st m_max;
-        bool        m_applyToSubGridAreas;
-    };
-
-private:
-    std::vector<CellRange> m_includeRanges;
-    std::vector<CellRange> m_excludeRanges;
-};
-
-//--------------------------------------------------------------------------------------------------
-///
-//--------------------------------------------------------------------------------------------------
 class CellFaceVisibilityFilter
 {
 public:
